Final Update main.py
Improvements made for the Code and the Console output:

Code:
Included descriptions and comments to the code for clarity.
Output:
Enhanced the output. Included headers and separators. Each column width is determined dynamically by getting the maximum  width of the content and ensured neat alignment. Made use of ljust() method to present the data in a visually organised way.

Color coding is done by making use of the "colorama" module for improved readability and visual appeal.